+++ nekro_agent/adapters/onebot_v11/matchers/guard.py
@@ -4,7 +4,7 @@ from nonebot.adapters import Bot, Message
 from nonebot.adapters.onebot.v11 import GroupMessageEvent, MessageEvent
 from nonebot.matcher import Matcher
 
-from nekro_agent.adapters.nonebot.tools.onebot_util import (
+from nekro_agent.adapters.onebot_v11.tools.onebot_util import (
     get_chat_info_old,
     get_user_name,
 )
