+++ nekro_agent/adapters/onebot_v11/tools/convertor.py
@@ -10,7 +10,7 @@ from nonebot.adapters.onebot.v11 import (
 )
 
 from nekro_agent.adapters.interface.base import BaseAdapter
-from nekro_agent.adapters.nonebot.tools.onebot_util import get_user_group_card_name
+from nekro_agent.adapters.onebot_v11.tools.onebot_util import get_user_group_card_name
 from nekro_agent.core.config import config
 from nekro_agent.core.logger import logger
 from nekro_agent.core.os_env import NAPCAT_TEMPFILE_DIR
