+++ plugins/builtin/basic.py
@@ -8,7 +8,7 @@ import aiofiles
 import magic
 from pydantic import Field
 
-from nekro_agent.adapters.nonebot.tools import user
+from nekro_agent.adapters.onebot_v11.tools import user
 from nekro_agent.api import core, message
 from nekro_agent.api.plugin import ConfigBase, NekroPlugin, SandboxMethodType
 from nekro_agent.api.schemas import AgentCtx
